requirements.txt
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
bertopic==0.16.4
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
cymem==2.0.11
debugpy==1.8.12
decorator==5.1.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
filelock==3.16.1
fsspec==2024.12.0
gensim==4.3.3
hdbscan==0.8.40
huggingface-hub==0.27.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
preshed==3.0.9
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==3.3.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.3.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.2
